+++ Changes.md
@@ -11,7 +11,8 @@ Improvements
 ------------
 
 - Viewer : Added visualisation support for Arnold shader networks connected to light gobos (#3667).
-- Catalogue : Added column to help identify the nature of each image (#3646)
+- Render nodes : Added context variables to image metadata (#3646). Currently only a subset of variable types is supported, depending on the renderer used.
+- Catalogue : Added customisable columns to display image information (#3646).
 
 Fixes
 -----